Ignore the last commit. This one is more complete.
Working on weapon shop. Added a sidebar for info. This was done by putting everything from the before commit into a panel rather than a frame. A new frame was created to encapsulate this and add a new panel on the side.

It references Game.SideBar for sidebar panel.

THIS COMMIT ADDS THE ACTUAL WEAPONS AND DEALS WITH SPACING